feat(cms): add featured item to payload CMS for work items
